moving logging.py _logging.py
